zcash_protocol: move TargetHeight into the consensus module
`TargetHeight` wraps a `BlockHeight` to represent the next chain tip and
carries no dependency on `zcash_client_backend`, so it belongs in the
lower `zcash_protocol` crate alongside `BlockHeight`. Move the definition
there and re-export it from `zcash_client_backend::data_api::wallet` so
existing call sites are unaffected.

This also lets the no_std `zcash_pool_migration_backend` state machine,
which depends on `zcash_protocol` but not on the optional
`zcash_client_backend`, adopt `TargetHeight` for its target-height
parameters.

Addresses review feedback on PR #2710 (r3632920093).
