+++ components/zcash_protocol/CHANGELOG.md
@@ -13,6 +13,9 @@ workspace.
 ### Added
 - `zcash_protocol::consensus::SECONDS_PER_BLOCK`
 - `zcash_protocol::consensus::BLOCKS_PER_HOUR`
+- `zcash_protocol::consensus::TargetHeight`, a wrapper around `BlockHeight`
+  representing the next chain tip. Previously defined in
+  `zcash_client_backend::data_api::wallet`.
 - `zcash_protocol::constants::MAX_BLOCK_BYTES`, the Zcash consensus maximum
   block size and therefore the maximum size of any single transaction.
 - `zcash_protocol::value::Zatoshis::{write, read, to_u64_le_bytes}`, a canonical
